Animesh_temp/TarkShaastra/requirements.txt
flask
flask-sqlalchemy
flask-login
appwrite
python-dotenv
requests
bcrypt
